Add details and correct spelling and wording
I've added details for what we consider a Code4PR project and we consider a community project.
I've also corrected some spelling mistakes and improved some of the wording in the document.

This is still very much an initial/rough draft. We need tons of feedback on this from the community.